Fix two minor errors
A bug involving mismatched `str` and `bytes` appears when
`allow_black_foreground` is false on Python 3; removing `.encode` from
the `black()` function resolves this.

Additionally, `legit --config` starts an editor with the path to the
config in the buffer, instead of opening the config itself. The
`filename=` kwarg needs to be used to correctly open the file.