fix: Control errors when working on empty repo
When we try to use legit over an initialized
but empty repo it breaks with an stacktrace.

Now we control that the repo has at least one head
and for 'undo' command also that we have at least
two commits on the branch. This is needed because
'git reset' cannot work on empty repo or undo
'root' commit.